[DSL/SMT] WIP HO quantifier elimination as `simproc`
This seems to work when there is a single quantification over `State`,
including when this quantification is nested in the expression
(which didn't work in the previous implementation).

HOWEVER, it still does not work when there are multiple states
being quantifier over, e.g. when an action makes multiple calls.
I suspect the solution will be a closer integration between
`State_eq_push_left` and `pushEqInvolvingLeft`.