[SMT] WIP HO quantifier elimination
This works towards eliminating HO quantification over `State`, but
still does not work when the action makes multiple calls (as in
commit `371be65`).

It appears that

```Lean
(pushEqInvolvingLeft lastQ) |> Simp.andThen (State_exists_push_right lastQ)
```

is not the right way to do this, since it leads to an infinite loop.
I wonder if instead we should run `Simp.main` twice (once for each
operation) and then chain the proofs together. Perhaps that will work.