[DSL] `assumption` (axiom) definition
It's a bit annoying that `axiom` is already a keyword in `command`
scope, so we can't really reuse it.

An `assumption` declaration checks that the variables referred to
in its body are not marked `mutable`. This check IS NOT IMPLEMENTED
for non-DSL `assumption`s, which might cause errors in the future.