[DSL] Access pre-state in `after_init`; remove `lang1`
This gets rid of an annoying source of code duplication and
solves #62.

An unfortunate side-effect of this commit is that we can no
longer use `_` to quantify in `after_init`, e.g.:

```lean
n_requesting _ := False
```

must now be:

```lean
n_requesting N := False
```

This commit fixes some, but NOT ALL the case studies to address
this.