[DSL] Derive `Nonempty`; `call act` now works instead of `call !act`
Deriving `Nonempty` for the `State` type doesn't work if there are
typeclass parameters (e.g. `[DecidableEq node]`) passed to it, so
we now get rid of those (only for the `State` type -- actions still
retain them, matching the previous behaviour.)

(We derive `Nonempty` to enable some simplifications when quantifying
over the `State` type.)

This seems to have broken the `notation` we used for `call !act`.
To fix this, we now define `act` (previously `act.fn`) and make its
arguments implicit. This means we can now use `call act` instead of
the previous cumbersome and hacky syntax of `call !act`.