[SMT/DSL] Working `#check_invariants` and `trace` with `Wlp` semantics
I will probably have to look into the performance of `simplify_all`
and `sdestruct_hyps` (I have had to manually increase heartbeats),
but this works!

I've also added a small test that `bmc_sat` should succeed as a
tactic even if the goal is trivial. (This is similar to the test in
`SolveClauseTrivialSuceed` and the bug had the same root cause.)