refactor: neatly separate out syntax of spec DSL
Syntax and semantics was all mixed up, which made it quite difficult
to see what's actually in the language. There's still some refactoring
to do in `Specification/Lang.lean` (it's way too long), but at least
now we have `Specification/Syntax.lean` as a reference that we
can build documentation on top of.