fix: decrease simp set for action simplification
This is still a major problem (bad performance). In the future, we
need to stage the order we do simplification in a lot more
carefully to reduce the time taken with quantifier hoisting &
elimination.

I wonder whether there's a clever way of setting the rule priorities
to achieve this staging effect that we want.