feat: ensure child module's state is immutable within parent module
I've kept the code for `passthrough` (formerly called `module`
mutability), as I suspect that will prove useful in the future.

For future reference: I am now intentionally conflating  Ivy's notions
of `module` and `isolate` — in Ivy, modules _can_ modify included
modules' state, but isolates cannot modify other isolates' state.

Basically, we are trying to make it such that every Veil module
is also an isolate (in the Ivy sense), but it might turn out that
eventually we need to make the same distinction as Ivy.