chore: remove IO automata terminology; LTS should be enough for us
Labelled Transition Systems are similar to IO Automata, but do not
have the same composition principles. Since we're implementing
modularity at the level of the Veil imperative DSL for now and
it's unclear how OOP-style composition and IOA-style composition
should interact/interplay, let's ditch the latter for now.

We can add it back later (as a composition principle for
transition-style specifications) if needed.