wip: use `applyAttributes` to allow in-time `simp` attribute registration
Found that `applyAttributesAt` will affect whether a definition can be
added to some `simp` attribute in time, which can in turn affect WP
reuse. So a cheap fix is to use `applyAttributes` for those
definitions/theorems.

Not sure if this is a good idea. Maybe this will be reverted later.