fix: add some basic instances to `invSimp`
Previously there is a commit that adds almost the same things to
`substateSimp`. This commit resolves the issue where unsimplified
`id` and `instIsSubStateOfRefl` blocks SMT query translation when
using `LocalRProp`; the previous one does not work since when using
`LocalRProp`, `substateSimp` is not enabled.

However, it is unclear whether the previous commit is not necessary,
since the same issue might happen when ghost relations are used inside
actions and we want to simplify them during WP generation. So let's keep
both for now.