fix: record definition actual params as `Parameter`
This standardises the representation - all parameters are now fully
captured in our internal representation of modules and definitions,
whereas previously the 'actual' definition parameters (i.e. those
directly specified by the user) where stored as `Syntax` of the
binders.

This is in preparation for removing magic constants from
`proveLocalityForStatePredicateCore` and `replaceLocalRProp`.